move MeetingToken alias to jwt
