+++ src/client/html/index.html
@@ -107,13 +107,13 @@
             class="mic-off"
             aria-label="Toggle microphone"
           ></button>
-          <button id="muteAll" aria-label="Mute all" class="hidden">Mute all</button>
+          <button id="muteAll" aria-label="Mute all" class="invisible">Mute all</button>
           <button id="leave">Leave game</button>
           <div class="right">
             <button id="invite" aria-label="Copy invite link">
               Copy invite link
             </button>
-            <button id="restart" aria-label="Restart" class="hidden">
+            <button id="restart" aria-label="Restart" class="invisible">
               Restart
             </button>
           </div>
